broke out message feed card into separate dir
