+++ react-app/src/components/MessageFeed/MessageFeed.js
@@ -1,9 +1,11 @@
 import React, { useEffect } from "react";
 import { useSelector, useDispatch } from "react-redux";
 
-import { loadAllUserMessages } from "../../store/messages";
-import MessageFeedCard from "./MessageFeedCard";
+import MessageFeedCard from "../MessageFeedCard";
 import Title from "../Title";
+import CreateMessage from "../CreateMessage";
+
+import { loadAllUserMessages } from "../../store/messages";
 
 import styles from "./MessageFeed.module.css";
 // className={styles. }
@@ -17,9 +19,10 @@ export default function MessageFeed({ message }) {
     }, [dispatch, user.id]);
 
     return (
-        <>
-            <Title className={styles.title} title="All Messages" />
+        <div className={styles.wrapper}>
+            <Title title="All Messages" />
             <MessageFeedCard message={message} />
-        </>
+            <CreateMessage />
+        </div>
     );
 }
