+++ react-app/src/components/MessageFeed/MessageFeed.module.css
@@ -1,12 +1,4 @@
-.title {
+.wrapper {
     display: flex;
-    justify-content: center;
-    font-size: 60px;
-    font-weight: 600;
-    margin: 0;
-    text-shadow: 1px 1px 1px darkgreen, 2px 2px 1px #0bda51;
-}
-
-.messages {
-    margin: 0 8px;
+    flex-direction: column;
 }
