+++ react-app/src/components/MessageFeedCard/index.js
@@ -1,0 +1,3 @@
+import MessageFeedCard from "./MessageFeedCard";
+
+export default MessageFeedCard;
